Add better clear and fix app container deepcopy
When the container is copied by scrapy, the file object is copied as well.
It is then read multiple times, because the config is loaded per instance
of a container object. Because the file object is never reset, an empty
config is loaded.

To fix it, the file is opened each time the config is read.